Make sure the arg{max,min} backend test cases have int64 as output type

+++ onnx/backend/test/case/node/argmin.py
@@ -14,7 +14,7 @@ def argmin_use_numpy(data, axis=0, keepdims=1):  # type: (np.ndarray, int, int) 
     result = np.argmin(data, axis=axis)
     if (keepdims == 1):
         result = np.expand_dims(result, axis)
-    return result
+    return result.astype(np.int64)
 
 
 class ArgMin(Base):
